requirements.txt
absl-py==0.7.0
aniso8601==3.0.2
appnope==0.1.0
asgiref==3.2.3
astor==0.7.1
astroid==1.6.5
atomicwrites==1.3.0
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.6.3
bleach==2.1.4
bs4==0.0.1
certifi==2018.8.24
chardet==3.0.4
click==6.7
cloudpickle==0.7.0
codecov==2.0.15
colorama==0.4.1
coloredlogs==10.0
coverage==4.5.4
cycler==0.10.0
Cython==0.29.6
decorator==4.4.1
dj-database-url==0.5.0
Django==3.0
djangorestframework==3.10.3
google-pasta==0.1.7
grpcio==1.24.1
gunicorn==20.0.4
h5py==2.10.0
html5lib==1.0.1
humanfriendly==4.17
idna==2.8
imageio==2.6.1
importlib-metadata==0.23
imutils==0.5.3
isort==4.3.4
joblib==0.14.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.3.1
Markdown==3.1.1
matplotlib==3.1.1
mccabe==0.6.1
more-itertools==7.2.0
networkx==2.4
numpy==1.17.4
opencv-python==4.1.1.26
packaging==19.2
pdfminer3k==1.3.1
Pillow==6.2.0
pluggy==0.13.0
ply==3.11
protobuf==3.10.0
psycopg2==2.8.4
py==1.8.0
pyparsing==2.4.2
pyreadline==2.1
pytesseract==0.3.0
pytest==5.2.1
pytest-cov==2.8.1
python-dateutil==2.8.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.1.2
requests==2.22.0
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.3.1
six==1.11.0
sklearn==0.0
sqlparse==0.3.0
termcolor==1.1.0
toolz==0.9.0
tqdm==4.36.1
urllib3==1.25.6
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
whitenoise==5.0.1
wincertstore==0.2
zipp==0.6.0
